server/pyproject.toml
[project]
name = "eduforge-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "google-genai>=1.43.0",
    "livekit-agents>=1.0.0",
    "livekit-api>=1.0.0",
    "asyncpg>=0.30.0",
    "redis[hiredis]>=5.0.0",
    "pycrdt-websocket>=0.15.0",
    "pycrdt>=0.12.0",
    "pydantic-settings>=2.7.0",
    "structlog>=24.0.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
